combine reducers into a single case statement
