+++ src/client/modules/Reminders/Reminders.jsx
@@ -2,8 +2,68 @@ import React from 'react'
 import { useParams, Redirect } from 'react-router-dom'
 import { connect } from 'react-redux'
 
-import { RemindersLists } from '.'
+import { SPACING, MEDIA_QUERIES } from '@govuk-react/constants'
+import styled from 'styled-components'
+import { Link } from 'govuk-react'
+
 import { state2props } from './state'
+
+import FooterLink from './FooterLink'
+import { DefaultLayout } from '../../components'
+import RemindersMenu from './RemindersMenu'
+
+import InvestmentsEstimatedLandDatesList from './InvestmentsEstimatedLandDatesList'
+import InvestmentsNoRecentInteractionsList from './InvestmentsNoRecentInteractionsList'
+import InvestmentsOutstandingPropositionsList from './InvestmentsOutstandingPropositionsList'
+import ExportsNoRecentInteractionsList from './ExportsNoRecentInteractionsList'
+import ExportsNewInteractionsList from './ExportsNewInteractionsList'
+import MyTasksDueDateApproachingList from './MyTasksDueDateApproachingList'
+
+import urls from '../../../lib/urls'
+import {
+  reminderTypeToLabel,
+  INVESTMENTS_ESTIMATED_LAND_DATES,
+  INVESTMENTS_NO_RECENT_INTERACTIONS,
+  INVESTMENTS_OUTSTANDING_PROPOSITIONS,
+  COMPANIES_NO_RECENT_INTERACTIONS,
+  COMPANIES_NEW_INTERACTIONS,
+  MY_TASKS_DUE_DATE_APPROACHING,
+  TASK_ASSIGNED_TO_ME_FROM_OTHERS,
+} from './constants'
+import TaskAssignedToMeFromOthersList from './TaskAssignedToMeFromOthersList'
+
+const Container = styled('div')({
+  [MEDIA_QUERIES.DESKTOP]: {
+    display: 'flex',
+  },
+})
+
+const MenuContainer = styled('div')({
+  width: '100%',
+  [MEDIA_QUERIES.DESKTOP]: {
+    width: 'calc(33% - 20px)',
+    padding: SPACING.SCALE_2,
+  },
+})
+
+const SettingsLink = styled(Link)({
+  display: 'block',
+  marginTop: SPACING.SCALE_5,
+  marginBottom: SPACING.SCALE_3,
+  [MEDIA_QUERIES.TABLET]: {
+    display: 'none',
+  },
+  [MEDIA_QUERIES.DESKTOP]: {
+    display: 'none',
+  },
+})
+
+const ListContainer = styled('div')({
+  width: '100%',
+  [MEDIA_QUERIES.DESKTOP]: {
+    width: '67%',
+  },
+})
 
 export const Reminders = ({ defaultUrl }) => {
   const { reminderType } = useParams()
@@ -11,7 +71,63 @@ export const Reminders = ({ defaultUrl }) => {
   if (!reminderType) {
     return <Redirect to={{ pathname: defaultUrl }} />
   }
-  return <RemindersLists reminderType={reminderType} />
+  const subject = reminderTypeToLabel[reminderType]
+  return (
+    <DefaultLayout
+      pageTitle={`Reminders - ${subject}`}
+      heading="Reminders"
+      subheading={subject}
+      breadcrumbs={[
+        { link: urls.dashboard.index(), text: 'Home' },
+        { link: urls.reminders.index(), text: 'Reminders' },
+        { text: subject },
+      ]}
+    >
+      <>
+        <Container>
+          <MenuContainer>
+            <RemindersMenu />
+            {reminderType !== INVESTMENTS_OUTSTANDING_PROPOSITIONS && (
+              <SettingsLink
+                data-test="reminders-settings-link"
+                href="/reminders/settings"
+              >
+                Reminders settings
+              </SettingsLink>
+            )}
+          </MenuContainer>
+          <ListContainer>
+            {reminderType === INVESTMENTS_ESTIMATED_LAND_DATES && (
+              <InvestmentsEstimatedLandDatesList />
+            )}
+            {reminderType === INVESTMENTS_NO_RECENT_INTERACTIONS && (
+              <InvestmentsNoRecentInteractionsList />
+            )}
+            {reminderType === INVESTMENTS_OUTSTANDING_PROPOSITIONS && (
+              <InvestmentsOutstandingPropositionsList />
+            )}
+            {reminderType === COMPANIES_NO_RECENT_INTERACTIONS && (
+              <ExportsNoRecentInteractionsList />
+            )}
+            {reminderType === COMPANIES_NEW_INTERACTIONS && (
+              <ExportsNewInteractionsList />
+            )}
+            {reminderType === MY_TASKS_DUE_DATE_APPROACHING && (
+              <MyTasksDueDateApproachingList />
+            )}
+            {reminderType === TASK_ASSIGNED_TO_ME_FROM_OTHERS && (
+              <TaskAssignedToMeFromOthersList />
+            )}
+          </ListContainer>
+        </Container>
+        <FooterLink
+          headingText="Need Help?"
+          linkUrl={urls.external.reminderAndSettings}
+          linkText="guidance on reminders and email notifications"
+        />
+      </>
+    </DefaultLayout>
+  )
 }
 
 export default connect(state2props)(Reminders)
