+++ src/client/modules/Reminders/tasks.js
@@ -1,59 +1,55 @@
 import { getPageOffset } from '../../utils/pagination'
 import { apiProxyAxios } from '../../components/Task/utils'
 
+const getReminders = (sortby = '-created_on', page = 1, limit = 10, endpoint) =>
+  apiProxyAxios
+    .get(`/v4/reminder/${endpoint}`, {
+      params: { sortby, limit, offset: getPageOffset({ page, limit }) },
+    })
+    .then(({ data }) => data)
+
+const getNextReminder = (
+  sortby = '-created_on',
+  page = 1,
+  limit = 10,
+  endpoint
+) =>
+  apiProxyAxios
+    .get(`/v4/reminder/${endpoint}`, {
+      params: {
+        sortby,
+        limit: 1,
+        offset: Math.max(getPageOffset({ page: page + 1, limit }) - 1, 1),
+      },
+    })
+    .then(({ data }) => data.results)
+
 // *************************** Investment lists ***************************
 
 export const getEstimatedLandDateReminders = ({
   sortby = '-created_on',
   page = 1,
   limit = 10,
-} = {}) =>
-  apiProxyAxios
-    .get('/v4/reminder/estimated-land-date', {
-      params: { sortby, limit, offset: getPageOffset({ page, limit }) },
-    })
-    .then(({ data }) => data)
+} = {}) => getReminders(sortby, page, limit, 'estimated-land-date')
 
 export const getNextEstimatedLandDateReminder = ({
   sortby = '-created_on',
   page = 1,
   limit = 10,
-} = {}) =>
-  apiProxyAxios
-    .get('/v4/reminder/estimated-land-date', {
-      params: {
-        sortby,
-        limit: 1,
-        offset: Math.max(getPageOffset({ page: page + 1, limit }) - 1, 1),
-      },
-    })
-    .then(({ data }) => data.results)
+} = {}) => getNextReminder(sortby, page, limit, 'estimated-land-date')
 
 export const getNoRecentInteractionReminders = ({
   sortby = '-created_on',
   page = 1,
   limit = 10,
-} = {}) =>
-  apiProxyAxios
-    .get('/v4/reminder/no-recent-investment-interaction', {
-      params: { sortby, limit, offset: getPageOffset({ page, limit }) },
-    })
-    .then(({ data }) => data)
+} = {}) => getReminders(sortby, page, limit, 'no-recent-investment-interaction')
 
 export const getNextNoRecentInteractionReminder = ({
   sortby = '-created_on',
   page = 1,
   limit = 10,
 } = {}) =>
-  apiProxyAxios
-    .get('/v4/reminder/no-recent-investment-interaction', {
-      params: {
-        sortby,
-        limit: 1,
-        offset: Math.max(getPageOffset({ page: page + 1, limit }) - 1, 1),
-      },
-    })
-    .then(({ data }) => data.results)
+  getNextReminder(sortby, page, limit, 'no-recent-investment-interaction')
 
 export const deleteEstimatedLandDateReminder = ({ id } = {}) =>
   apiProxyAxios.delete(`/v4/reminder/estimated-land-date/${id}`)
@@ -90,27 +86,13 @@ export const getExportsNoRecentInteractionReminders = ({
   sortby = '-created_on',
   page = 1,
   limit = 10,
-} = {}) =>
-  apiProxyAxios
-    .get('/v4/reminder/no-recent-export-interaction', {
-      params: { sortby, limit, offset: getPageOffset({ page, limit }) },
-    })
-    .then(({ data }) => data)
+} = {}) => getReminders(sortby, page, limit, 'no-recent-export-interaction')
 
 export const getNextExportNoRecentInteractionReminder = ({
   sortby = '-created_on',
   page = 1,
   limit = 10,
-} = {}) =>
-  apiProxyAxios
-    .get('/v4/reminder/no-recent-export-interaction', {
-      params: {
-        sortby,
-        limit: 1,
-        offset: Math.max(getPageOffset({ page: page + 1, limit }) - 1, 1),
-      },
-    })
-    .then(({ data }) => data.results)
+} = {}) => getNextReminder(sortby, page, limit, 'no-recent-export-interaction')
 
 export const deleteExportNoRecentInteractionReminder = ({ id } = {}) =>
   apiProxyAxios.delete(`/v4/reminder/no-recent-export-interaction/${id}`)
@@ -119,27 +101,13 @@ export const getExportsNewInteractionReminders = ({
   sortby = '-created_on',
   page = 1,
   limit = 10,
-} = {}) =>
-  apiProxyAxios
-    .get('/v4/reminder/new-export-interaction', {
-      params: { sortby, limit, offset: getPageOffset({ page, limit }) },
-    })
-    .then(({ data }) => data)
+} = {}) => getReminders(sortby, page, limit, 'new-export-interaction')
 
 export const getNextExportsNewInteractionReminder = ({
   sortby = '-created_on',
   page = 1,
   limit = 10,
-} = {}) =>
-  apiProxyAxios
-    .get('/v4/reminder/new-export-interaction', {
-      params: {
-        sortby,
-        limit: 1,
-        offset: Math.max(getPageOffset({ page: page + 1, limit }) - 1, 1),
-      },
-    })
-    .then(({ data }) => data.results)
+} = {}) => getNextReminder(sortby, page, limit, 'new-export-interaction')
 
 export const deleteExportNewInteractionReminder = ({ id } = {}) =>
   apiProxyAxios.delete(`/v4/reminder/new-export-interaction/${id}`)
@@ -150,27 +118,13 @@ export const getMyTasksDueDateApproachingReminders = ({
   sortby = '-created_on',
   page = 1,
   limit = 10,
-} = {}) =>
-  apiProxyAxios
-    .get('/v4/reminder/my-tasks-due-date-approaching', {
-      params: { sortby, limit, offset: getPageOffset({ page, limit }) },
-    })
-    .then(({ data }) => data)
+} = {}) => getReminders(sortby, page, limit, 'my-tasks-due-date-approaching')
 
 export const getMyTasksNextDueDateApproachingReminder = ({
   sortby = '-created_on',
   page = 1,
   limit = 10,
-} = {}) =>
-  apiProxyAxios
-    .get('/v4/reminder/my-tasks-due-date-approaching', {
-      params: {
-        sortby,
-        limit: 1,
-        offset: Math.max(getPageOffset({ page: page + 1, limit }) - 1, 1),
-      },
-    })
-    .then(({ data }) => data.results)
+} = {}) => getNextReminder(sortby, page, limit, 'my-tasks-due-date-approaching')
 
 export const deleteMyTasksDueDateApproachingReminder = ({ id } = {}) =>
   apiProxyAxios.delete(`/v4/reminder/my-tasks-due-date-approaching/${id}`)
